requirements.txt
Adafruit_DHT ~= 1.4.0
python-telegram-bot ~= 13.4.1
